don't add course projects on assignment
